+++ crates/fuel-core/src/graphql_api/api_service.rs
@@ -15,6 +15,7 @@ use crate::{
         view_extension::ViewExtension,
         Config,
     },
+    graphql_api::QUERY_COSTS,
     schema::{
         CoreSchema,
         CoreSchemaBuilder,
@@ -220,6 +221,7 @@ where
     OnChain::LatestView: OnChainDatabase,
     OffChain::LatestView: OffChainDatabase,
 {
+    initialize_query_costs(config.clone());
     let network_addr = config.config.addr;
     let combined_read_database =
         ReadDatabase::new(genesis_block_height, on_database, off_database);
@@ -294,6 +296,10 @@ where
         GraphqlService { bound_address },
         ServerParams { router, listener },
     ))
+}
+
+fn initialize_query_costs(config: Config) {
+    QUERY_COSTS.get_or_init(|| config.config.costs);
 }
 
 async fn graphql_playground() -> impl IntoResponse {