+++ crates/fuel-core/src/service/config.rs
@@ -147,6 +147,7 @@ impl Config {
                 request_body_bytes_limit: 16 * 1024 * 1024,
                 query_log_threshold_time: Duration::from_secs(2),
                 api_request_timeout: Duration::from_secs(60),
+                costs: Default::default(),
             },
             combined_db_config,
             continue_on_error: false,